# api-documentation.yaml
openapi: 3.0.0
info:
  title: Party Picker API
  version: 1.0.0
...
security:
  - BearerAuth: []
  - BasicAuth: []
paths:
...
  /services:
    get:
      summary: Get services
      tags:
        - Services
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: The page number to retrieve
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: The number of services per page
        - in: query
          name: category
          schema:
            type: string
          description: The category of the services
        - in: query
          name: minPrice
          schema:
            type: number
            minimum: 0
          description: The minimum price of the services
        - in: query
          name: maxPrice
          schema:
            type: number
            minimum: 1
          description: The maximum price of the services
        - in: query
          name: attendees
          schema:
            type: integer
            minimum: 0
          description: The number of attendees that the services can accommodate
        - in: query
          name: districts
          schema:
            type: array
            items:
              type: string
          description: The districts where the services are located
        - in: query
          name: likedBy
          schema:
            type: integer
          description: The ID of the user that liked the services
        - in: query
          name: sortType
          schema:
            type: string
            enum:
              - RATING_DESC
              - REVIEW_COUNT_DESC
              - POPULARITY_DESC
        - in: query
          name: search
          schema:
            type: string
          description: The search query
        - in: query
          name: availableOn
          schema:
            type: string
            format: date-time
          description: The date in which the services need to be available
        - in: query
          name: createdBy
          schema:
            type: integer
          description: The ID of the user that created the services
        - in: query
          name: event
          schema:
            type: integer
          description: The ID of the event that the services requested are related to
        - in: query
          name: includeDeleted
          schema:
            type: boolean
          description: Whether to include deleted services
        - in: query
          name: recommendedForEvent
          schema:
            type: integer
          description: The ID of the event that the services requested are recommended for. Mutually exclusive with the other parameters, except for `page` and `pageSize`
        - in: query
          name: recommendedForOffering
          schema:
            type: integer
          description: The ID of the offering that the services requested are recommended for. Mutually exclusive with the other parameters, except for `page` and `pageSize`
      responses:
        "200":
          description: The services were successfully retrieved
          headers:
            first:
              $ref: "#/components/headers/first"
            last:
              $ref: "#/components/headers/last"
            next:
              $ref: "#/components/headers/next"
            prev:
              $ref: "#/components/headers/prev"
          content:
            application/vnd.services.v1+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
        "400":
          description: Invalid request.
    post:
      summary: Create a service
      tags:
        - Services
      requestBody:
        description: The details of the new service
        required: true
        content:
          application/vnd.service.v1+json:
            schema:
              $ref: "#/components/schemas/Service"
      responses:
        "201":
          description: The service was successfully created
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
  /services/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 0
        description: The ID of the service
    get:
      summary: Get a service by ID
      tags:
        - Services
      responses:
        "200":
          description: The service was successfully retrieved
          content:
            application/vnd.service.v1+json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: The service was not found
    put:
      summary: Update a service
      tags:
        - Services
      requestBody:
        description: The new service details
        required: true
        content:
          application/vnd.service.v1+json:
            schema:
              $ref: "#/components/schemas/Service"
      responses:
        "200":
          description: The service was successfully updated
          content:
            application/vnd.service.v1+json:
              schema:
                $ref: "#/components/schemas/Service"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The service was not found
    delete:
      summary: Delete a service
      tags:
        - Services
      responses:
        "204":
          description: The service was successfully deleted
        "404":
          description: The service was not found
  /services/{id}/likes:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 0
        description: The ID of the service
    post:
      summary: Like a service
      tags:
        - Services
      requestBody:
        description: The details of the new like
        required: true
        content:
          application/vnd.like.v1+json:
            schema:
              $ref: "#/components/schemas/Like"
      responses:
        "201":
          description: The service was successfully liked
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The service was not found
  /services/{id}/likes/{userId}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 0
        description: The ID of the service
      - in: path
        name: userId
        required: true
        schema:
          type: integer
          minimum: 0
        description: The ID of the user
    get:
      summary: Get a like by ID
      tags:
        - Services
      responses:
        "200":
          description: The like was successfully retrieved
          content:
            application/vnd.like.v1+json:
              schema:
                $ref: "#/components/schemas/Like"
        "404":
          description: The like was not found
    delete:
      summary: Unlike a service
      tags:
        - Services
      responses:
        "204":
          description: The service was successfully unliked
        "404":
          description: The service or user was not found
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Service:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: The ID of the service
        name:
          type: string
          minLength: 4
          maxLength: 32
          description: The name of the service
        description:
          type: string
          description: The description of the service
        category:
          type: string
          description: The category of the service
        minPrice:
          type: number
          description: The minimum price of the service
        maxPrice:
          type: number
          description: The maximum price of the service
        priceType:
          type: string
          description: The price type of the service
        maxGuests:
          type: integer
          minimum: 0
          description: The maximum number of guests that the service can accommodate
        district:
          type: string
          description: The location of the service
        deleted:
          type: boolean
          readOnly: true
          description: Whether the service was deleted
        likes:
          type: integer
          readOnly: true
          description: The number of likes of the service
        rating:
          type: number
          readOnly: true
          description: The rating of the service
        self:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the full details of the service
        owner:
          type: string
          format: uri
          readOnly: true
          description: The URI of the user that created the service
        images:
          type: array
          items:
            type: string
            format: uri
          description: The URIs to the images of the service
        reviews:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the reviews of the service
    Like:
      type: object
      properties:
        self:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the full details of the like
        offering:
          type: string
          format: uri
          readOnly: true
          description: The URI of the service that was liked
        user:
          type: string
          format: uri
          description: The URI of the user that liked the service
        liked:
          type: boolean
          description: Whether the user liked the service
...
    FormErrors:
      type: array
      items:
        type: object
        properties:
          fieldName:
            type: string
            description: The name of the field that caused the error
          message:
            type: string
            description: The error message
...
  headers:
    first:
      schema:
        type: string
        format: uri
      description: The URI to the first page of the results
    last:
      schema:
        type: string
        format: uri
      description: The URI to the last page of the results
    next:
      schema:
        type: string
        format: uri
      description: The URI to the next page of the results, if there is one
    prev:
      schema:
        type: string
        format: uri
      description: The URI to the previous page of the results, if there is one
servers:
  - description: Party Picker API
    url: http://pawserver.it.itba.edu.ar/paw-2023b-04/api